requirements.txt
Django==4.2.8
whitenoise==6.4.0
python-dotenv==1.0.0
pytz==2024.1
sqlparse==0.4.4
asgiref==3.7.1
gunicorn==21.2.0
reportlab==4.0.7
Pillow==10.1.0
djangorestframework==3.14.0
django-filter==23.5
django-ratelimit==4.1.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
sentry-sdk==1.40.0
coverage==7.4.0
pytest-django==4.7.0